⚙️Huge changes
Refactored the entire code again! Now, creating and defining catalogs is much easier to understand and modify, thanks to a well-structured setup for the constructors, which are now in `parser.cpp`. Also changed how reading works for structs using the secondary offset

everything seems to be running smoothly now (for now xD)

still need to implement structs that use arrays in the secondary offset (markerSet is composed entirely of this and in Noun we have gfxStates that works like this too)